refactor : user module application layer 도입
