+++ src/main/java/com/meoguri/linkocean/controller/user/AuthController.java
@@ -16,8 +16,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.meoguri.linkocean.controller.user.dto.AuthRequest;
+import com.meoguri.linkocean.domain.user.application.OAuthAuthenticationService;
 import com.meoguri.linkocean.domain.user.model.OAuthType;
-import com.meoguri.linkocean.domain.user.service.OAuthAuthenticationService;
 
 import lombok.RequiredArgsConstructor;
