+++ src/main/java/com/meoguri/linkocean/domain/user/application/OAuthAuthenticationService.java
@@ -1,10 +1,11 @@
-package com.meoguri.linkocean.domain.user.service;
+package com.meoguri.linkocean.domain.user.application;
 
 import org.springframework.stereotype.Service;
 
 import com.meoguri.linkocean.configuration.security.jwt.JwtProvider;
 import com.meoguri.linkocean.domain.user.model.Email;
 import com.meoguri.linkocean.domain.user.model.OAuthType;
+import com.meoguri.linkocean.domain.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
