+++ src/main/java/com/meoguri/linkocean/domain/user/application/OAuthClient.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.user.service;
+package com.meoguri.linkocean.domain.user.application;
 
 import com.meoguri.linkocean.domain.user.model.Email;
