+++ src/main/java/com/meoguri/linkocean/domain/user/infrastructure/oauth/GoogleOAuthClient.java
@@ -14,8 +14,8 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.meoguri.linkocean.domain.user.application.OAuthClient;
 import com.meoguri.linkocean.domain.user.model.Email;
-import com.meoguri.linkocean.domain.user.service.OAuthClient;
 import com.meoguri.linkocean.exception.OAuthException;
 
 import lombok.RequiredArgsConstructor;
