+++ src/test/java/com/meoguri/linkocean/controller/user/AuthControllerTest.java
@@ -12,7 +12,7 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import com.meoguri.linkocean.controller.user.dto.AuthRequest;
 import com.meoguri.linkocean.domain.user.model.Email;
-import com.meoguri.linkocean.domain.user.service.OAuthClient;
+import com.meoguri.linkocean.domain.user.application.OAuthClient;
 import com.meoguri.linkocean.exception.OAuthException;
 import com.meoguri.linkocean.test.support.controller.BaseControllerTest;
