+++ src/test/java/com/meoguri/linkocean/test/restdocs/AuthRestDocsTest.java
@@ -13,8 +13,8 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import com.meoguri.linkocean.controller.user.AuthController;
 import com.meoguri.linkocean.controller.user.dto.AuthRequest;
+import com.meoguri.linkocean.domain.user.application.OAuthClient;
 import com.meoguri.linkocean.domain.user.model.Email;
-import com.meoguri.linkocean.domain.user.service.OAuthClient;
 import com.meoguri.linkocean.test.restdocs.support.RestDocsTestSupport;
 
 class AuthRestDocsTest extends RestDocsTestSupport {
